Sumazon Backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
bcrypt==4.1.2
    # via flask-bcrypt
blinker==1.7.0
    # via flask
click==8.1.7
    # via flask
flask==3.0.2
    # via
    #   -r requirements.in
    #   flask-bcrypt
    #   flask-cors
    #   flask-login
    #   flask-sqlalchemy
flask-bcrypt==1.0.1
    # via -r requirements.in
flask-cors==4.0.0
    # via -r requirements.in
flask-login==0.6.3
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
itsdangerous==2.1.2
    # via flask
jinja2==3.1.3
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
pyjwt==2.8.0
    # via -r requirements.in
sqlalchemy==2.0.25
    # via flask-sqlalchemy
typing-extensions==4.9.0
    # via sqlalchemy
werkzeug==3.0.1
    # via
    #   flask
    #   flask-login
